Fix mismatch of int64 vs int32 in new ramses particle format. Add unit tests for new format.

+++ pynbody/snapshot/ramses.py
@@ -484,7 +484,7 @@ class RamsesSnap(SimSnap):
         with open(self._filename+"/part_file_descriptor.txt") as f:
             self._particle_blocks = []
             self._particle_types = []
-            dtype_mappings = {'d': np.float64, 'i': np.int64, 'b': np.int8}
+            dtype_mappings = {'d': np.float64, 'i': np.int32, 'b': np.int8}
             self._translate_array_name = namemapper.AdaptiveNameMapper('ramses-name-mapping')
             for l in f:
                 if not l.startswith("#"):
